type bound procedures for gs and rf hamiltonian
Former-commit-id: fe7a423d7378e6946013bdbf8f3037a397176be4

+++ src/98_main/ioprof.F90
@@ -84,14 +84,9 @@ program ioprof
 #endif
 
  codename='IOPROF'//REPEAT(' ',17)
- call herald(codename,abinit_version,std_out)
+ call herald(codename, abinit_version, std_out)
 
- !verbose = .TRUE.
  ount = dev_null; if (verbose) ount = std_out
-
- if (nprocs>1) then
-   ABI_ERROR("not programmed for parallel execution, Run it in sequential")
- end if
 
  nargs = command_argument_count()
 
@@ -128,6 +123,7 @@ program ioprof
 
  case ("nc2fort")
    ! Converter: netcdf --> Fortran
+   if (my_rank /= master) goto 100
    ABI_CHECK(nargs == 3, "Usage: nc2form wfk_source wfk_dest")
    call get_command_argument(2, wfk_source)
    call get_command_argument(3, wfk_dest)
@@ -135,29 +131,33 @@ program ioprof
    if (file_exists(wfk_dest)) then
      ABI_ERROR(sjoin("Cannot overwrite:", wfk_dest, ". Remove file and rerun"))
    end if
-
-   if (my_rank == master) call wfk_nc2fort(wfk_source, wfk_dest)
+   call wfk_nc2fort(wfk_source, wfk_dest)
 
  case ("prof")
    ! Profile
    ABI_CHECK(nargs > 1, "Usage: bench wfk_path")
    call get_command_argument(2, wfk_path)
+   if (my_rank /= master) goto 100
+
    formeig = 0
    nband2read = 1500
    call wfk_prof(wfk_path, formeig, nband2read, comm)
    ! TO CREATE new file
    !call wfk_create_wfkfile(new_fname,hdr,iomode,formeig,Kvars,cwtimes,xmpi_comm_self)
 
+ !case ("nc_tests")
+
  case ("unittests")
    ! Unit tests.
+   if (my_rank /= master) goto 100
+
    do ii=1,count(hdr_fnames/=ABI_NOFILE)
 
      ! Read the header from an external netcdf file
      ! This trick is needed because the initialization of
      ! the header is *IMPOSSIBLE* if we don't have a full ABINIT input file!
-     !
      call hdr%from_fname(hdr_fnames(ii),fform,comm)
-     ABI_CHECK(fform/=0,"fform==0")
+     ABI_CHECK(fform /= 0, "fform==0")
 
      call hdr%echo(fform,4,unit=std_out)
 
@@ -165,13 +165,9 @@ program ioprof
        formeig = formeigs(feg)
        do io=1,size(io_modes)
          iomode = io_modes(io)
-
          new_fname = "NEW_WFK"
          if (iomode==IO_MODE_ETSF) new_fname = "NEW_WFK.nc"
-
-         if (file_exists(new_fname)) then
-           call delete_file(new_fname,ierr)
-         end if
+         if (file_exists(new_fname)) call delete_file(new_fname,ierr)
 
          ! TODO
          if (formeig==1 .and. iomode==IO_MODE_ETSF) then
@@ -182,10 +178,10 @@ program ioprof
          ABI_MALLOC(Kvars, (hdr%nkpt))
 
          if (my_rank == master) then
-           call wrtout(ount,"Calling wfk_create_wfkfile","COLL")
+           call wrtout(ount,"Calling wfk_create_wfkfile")
            call wfk_create_wfkfile(new_fname,hdr,iomode,formeig,Kvars,cwtimes,xmpi_comm_self)
            !call wfk_create_wfkfile(new_fname,hdr,IO_MODE_FORTRAN,formeig,Kvars,cwtimes,xmpi_comm_self)
-           call wrtout(ount,"Done wfk_create_wfkfile","COLL")
+           call wrtout(ount,"Done wfk_create_wfkfile")
          end if
 
          if (nprocs > 1) then
@@ -194,13 +190,11 @@ program ioprof
          end if
 
          method = 0
-
          call wrtout(std_out,sjoin("Checking file:",new_fname,", with iomode=",itoa(iomode)))
          call wfk_check_wfkfile(new_fname,hdr,iomode,method,formeig,Kvars,cwtimes,comm,ierr)
 
-         write(msg,'(2(a,i2),2(a,f8.2))')&
-         " Read with iomode: ",iomode,", nproc: ",nprocs,", cpu: ",cwtimes(1),", wall:",cwtimes(2)
-         call wrtout(ount,msg,"COLL")
+         write(msg,'(2(a,i0),2(a,f8.2))')" Read with iomode: ",iomode,", nproc: ",nprocs,", cpu: ",cwtimes(1),", wall:",cwtimes(2)
+         call wrtout(ount,msg)
 
          if (ierr/=0) then
            ABI_ERROR(sjoin("wfk_check_wfkfile returned ierr ",itoa(ierr)))
@@ -213,16 +207,11 @@ program ioprof
 
          if (check_iomode /= -100) then
            call wrtout(std_out,sjoin("Trying to read file:",new_fname,", with check_iomode=",itoa(check_iomode)))
-
            call wfk_check_wfkfile(new_fname,hdr,check_iomode,method,formeig,Kvars,cwtimes,comm,ierr)
 
-           write(msg,'(2(a,i2),2(a,f8.2))')&
-           " Read with iomode: ",iomode,", nproc: ",nprocs,", cpu: ",cwtimes(1),", wall:",cwtimes(2)
-           call wrtout(ount,msg,"COLL")
-
-           if (ierr/=0) then
-             ABI_ERROR(sjoin("wfk_check_wfkfile returned ierr ",itoa(ierr)))
-           end if
+           write(msg,'(2(a,i0),2(a,f8.2))')" Read with iomode: ",iomode,", nproc: ",nprocs,", cpu: ",cwtimes(1),", wall:",cwtimes(2)
+           call wrtout(ount,msg)
+           ABI_CHECK(ierr == 0, sjoin("wfk_check_wfkfile returned ierr ",itoa(ierr)))
          end if
 
          ABI_FREE(Kvars)
@@ -236,10 +225,9 @@ program ioprof
    ABI_ERROR(sjoin("Unknown command:", command))
  end select
 
- !call wrtout(std_out,ch10//" Analysis completed.","COLL")
+ !call wrtout(std_out,ch10//" Analysis completed.)
 
-!Writes information on file about the memory before ending mpi module, if memory profiling is enabled
- !ABI_MALLOC(nband, (2))
+ ! Writes information on file about the memory before ending mpi module, if memory profiling is enabled
  call abinit_doctor("__ioprof")
 
  100 call xmpi_end()
